rename snake cast to camel cast
